Impl: Add ability to play cards
Add the ability to play cards from a hand. The play method will return
the cards the player has selected and remove them from the players
hands. Also add an additional method to print cards using the playing
card unicode characters. This method is not used by default because the
default font-size is to small to read these cards. If a larger font-size
was used, this method is viable.